desktop/requirements.txt
bitarray==2.9.3
bitstring==4.1.4
dill==0.3.9
multiprocess==0.70.17
numpy==1.26.4
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
PyQt6==6.8.0
PyQt6-Qt6==6.8.1
PyQt6_sip==13.9.1
pyrtmp==0.3.1
ujson==5.10.0
